prints random seed in formatter summary
