+++ README.md
@@ -209,6 +209,7 @@ package main
 import (
 	"os"
 	"testing"
+	"time"
 
 	"github.com/DATA-DOG/godog"
 )
@@ -217,8 +218,9 @@ func TestMain(m *testing.M) {
 	status := godog.RunWithOptions("godogs", func(s *godog.Suite) {
 		FeatureContext(s)
 	}, godog.Options{
-		Format: "progress",
-		Paths:  []string{"features"},
+		Format:    "progress",
+		Paths:     []string{"features"},
+		Randomize: time.Now().UTC().UnixNano(), // randomize scenario execution order
 	})
 
 	if st := m.Run(); st > status {
